Don't look for dictionary item by name
SDF does not guarantee that names are unique, so searching by
name is not guaranteed to always work.
This was being used to build a dictionary of block ids. Better to
build the dictionary using the original block structure.
Since it is only currently used for finding mesh ids, that is all
that gets built.